requirements-dev.txt
-r requirements.txt

django-debug-toolbar
django-debug-toolbar-template-timings

fabric3

#tests ***
pytest-django==3.10.0
